+++ docs/changelog.rst
@@ -3,6 +3,20 @@ Changelog
 
 .. py:currentmodule:: ietfparse
 
+`Next Release`_
+---------------
+.. rubric:: Behavioural Change
+
+:func:`headers.parse_accept` used to fail with a :exc:`ValueError` when
+it encountered an invalid content type value in the header.  Now it skips
+the invalid value.  If you want the previous behaviour, then pass ``strict=True``.
+
+- Advertise support for Python 3.7-3.9, remove 3.4 & 3.5
+- Clarify that :func:`headers.parse_content_type` raises :exc:`ValueError`
+  when it encounters an invalid content type header
+- Skip unparseable content types in :func:`headers.parse_accept` unless
+  the new ``strict`` parameter is truthy
+
 `1.6.1`_ (26-Jan-2020)
 ----------------------
 - Fixed project URL metadata.